pkcs1: fixups for `DecodeRsaPublicKey` blanket impl
Changes the error type to `spki::Error`, which matches the error type
used in the `spki::DecodePublicKey` blanket impl:

https://docs.rs/spki/0.7.0/spki/trait.DecodePublicKey.html#impl-DecodePublicKey-for-T

Also removes the bounds on `DecodePublicKey` in the blanket impl of
`DecodeRsaPublicKey`, since they aren't actually used.

This is a breaking change but also a bugfix, as with the previous bounds
it's not possible to impl PKCS#1 support via these two blanket impls,
since the error types don't match. The plan is to merge this, publish
it, and yank the v0.7.0 release.