Handle cyclic dependencies
When cyclic dependencies are used for either doctests or integrations
tests (`x509-cert` depends on `rsa` which depends back to `der` in this
workspace), the compiler will complain about lack of trait
implementation because two crates will be pulled. One from this checkout
as well as one from crates.io.

This migrates the inclusion from local `path` in each Cargo.toml to a
global `patch.crates-io`.