pkcs7: deprecate in favor of `cms`
This commit adds a deprecation notice to README.md for the `pkcs7` crate
announcing `cms` as its official successor.

Notably `pkcs7` predates many of the other format-related crates and
uses entirely reference-based conventions which we migrated away from in
the `x509-cert` crate. `cms` uses the newer conventions, and is more
complete as a Cryptographic Message Syntax implementation.

We will not be accepting pull requests for `pkcs7` and are leaving its
source code for now only to aid in porting functionality not yet present
in the `cms` crate.

When the next breaking release of `der` occurs, we will delete the
`pkcs7` source code from the repository, leaving only the deprecation
notice to help assist people migrating.

Closes #1045.