der: add `SetOf(Vec)::insert(_ordered)`; deprecate `add`
Renames the insertion methods on `SetOf` and `SetOfVec` from `add` to
`insert_ordered`, deprecating the previous `add` method.

Also adds an `insert` method which pushes onto the underlying
`(Array)Vec` then calls `der_sort`. This should be relatively fast since
the underlying algorithm is insertion sort and operating on data which
is mostly in-order.

The name `insert` is more consistent with `BTreeSet`/`HashSet`.