Fix `repository` field in Cargo.toml
According to https://doc.rust-lang.org/cargo/reference/manifest.html#the-repository-field
`repository` is a source repository URL.
GitHub URLs with `tree` are not repository URLs
and are not usable as `git clone` arguments.

Path within the worktree
is encoded in `.cargo_vcs_info.json` of the crate,
so encoding it in `Cargo.toml` is not necessary.

I am automating downloading of crate sources
and invalid URLs require workarounds
to turn them into repository URLs that can be cloned.