Remove a `;`, to allow Rust to infer the type
`?` currently influences inference s.t. for `error()?` rustc infers `T = ()`.
However, it is quite confusing -- `?` is not supposed to influence inference,
it's just a conicidence because of the curent implementation. There is an idea
to change this behavior in future Rust versions, such that this code won't
compile without this change.