x509-cert: simplify CI (fewer cargo-hack combos)
Excludes several features from `cargo hack` to prevent a combinatorial
explosion, namely features which just activate other features in
dependencies:

- `arbitrary`
- `base64`
- `fingerprint`
- `hazmat`
- `sct`

The `arbitrary`, `hazmat`, and `sct` features are tested in isolation,
and finally the crate is tested with `--all-features` in both debug and
release modes.

This should significantly improve CI times.