const-oid: improve rustdoc on `Arc`
Note that we follow Mozilla conventions using maximum 32-bit arcs,
linking to a survey of various OID libraries and how they handle various
sized arcs.

Also notes we deliberatly do not support UUIDs-as-OIDs, since they're
not important (at least within the scope of what we currently implement)
to cryptographic use cases this library is intended for.