Helper function for BER to DER transcoding
Issue #779 has had some progress (PR #1321). However, this helper
function is intended to provide an escape hatch for limited support
of some cases of BER, while not waiting for general support or needing
API changes.

Specifically, the transcoding of occurrences of the constructed,
indefinite-length method into the constructed, definite-length method.
This is likely sufficient to address the examples in the wild, reported
in Issue #779 and elsewhere.

The goal is not to support all possible violations of DER that are still
following valid BER. Examples are non-canonical encodings of lengths
(which would require an alternative `Length`), or constructed string
types (which could be handled by further work on this function, but
usage may not be well motivated).