der: remove `IndefiniteLength` (#1807)
The original idea of this type was to have a type-safe distinction
between DER lengths (`Length`) and potentially indefinite BER lengths
(`IndefiniteLength`), but really what's needed is to make the `Length`
type support idefinite lengths when `EncodingRules::Ber` are in use.

Otherwise, we need to change `Header` to use `IndefiniteLength`, which
kind of defeats the point.