x509-cert: allow serialnumbers to be stored in a database
When storing serialnumbers in a database, a btree or similar structure,
the serial number could be required to implement Ord.
This is useful for a CRL database for example.

This is only possible if the Profile itself implements Ord. As this is
a very cheap requirement for the profile itself, this is now part of the
Profile trait.