der: add `Reader::peek`
Adds a method to the `Reader` trait for spawning a sub-reader that
starts at the current position but doesn't affect the parent reader's
state.

Implementing indefinite lengths will involve scanning large parts of the
message (possibly all the way to the end), so this is a building block
to making that work.