der: remove `Reader::read_value`
This was added in #1877 but the real value of that PR was the automatic
`read_nested` on `DecodeValue` impls. `Reader::read_value` isn't
actually necessary for that at all.

It may be useful in the future if we introduce a BER reader that can
handle decoding constructed e.g. strings from indefinite length
encoding, but since we don't actually have anything like that yet,
removing this avoids some unnecessary duplication and API surface for
the `Reader`.