mcf: add `Field` newtype
Also extracts a `fields` submodule.

The `Field` newtype is useful as a place to provide additional methods
for operating on individual fields, such as encoding/decoding various
flavors of Base64, or parsing `k=v,...` mappings.

The functionality to do that is not yet implemented, but having a type
where we can put such functionality is the starting place.