mcf: Base64 decoding support (#1918)
Adds a `Base64` enum with variants for `Bcrypt`, `Crypt`, and
`ShaCrypt`, i.e. the non-standard Base64 variants used in MCF hashes.

This is extracted from `password_hash::Encoding` which was straddling
the line between the PHC format and MCF with such a feature. That type
can now be removed in the next breaking release.