der: improve constructed bit handling (#1919)
Adds an internal decoding method which returns the constructed bit in
addition to the `Tag`.

This is currently used for validating the constructed bit at decode
time, however in future work it might be interesting to retain, either
on the `Tag` type (where we do store it for certain variants) or in the
`Header` type.