der: improved internal ref types
Changes `BytesRef` and `StringRef` from types with an explicit lifetime
which act as newtypes for inner `&'a [u8]` and `&'a str` types to being
newtypes for `[u8]` and `str`.

This requires a very small amount of `unsafe` casting, but makes it
possible to implement reference conversions entirely in terms of
standard traits like `AsRef`, `Borrow`, `Deref`, and `ToOwned` as
opposed to using our custom `RefToOwned` and OwnedToRef` traits.
It makes it possible to use e.g. `Cow` as well.

This is the first step towards moving the entire crate to this pattern
and eliminating the `OwnedToRef` and `RefToOwned` raits, replacing them
instead with the standard `AsRef` and `ToOwned` traits.