mcf: add `McfHashRef` (#1925)
Adds a reference type for MCF hashes which works in no-`alloc`
environments.

Some methods of `McfHash` are now directly delegated. It would be
possible to make a fully borrowed reference type that works like
`&McfHashRef`, which could be used with the standard
`AsRef`/`Borrow`/`Deref` `ToOwned` traits, but it would require unsafe
code, which this avoids for the time being.