base64ct, base32ct: proptests for decoder equivalence with base32/64
For the base64 case, this MR adds proptests to ensure that base64
and base64ct accept the same inputs when decoding, and produce the
same outputs when decoding them.

For the base32 case, that property isn't actually true.  Instead,
the proptest makes sure that whenever base32 and base32ct both
accept an input when decoding, they produce the same output.