phc: deprecate `PasswordHashString`
It was originally added because `PasswordHash` had a lifetime, so it
provided an owned alternative. But now `PasswordHash` is fully owned
(and also fully stack allocated, so no `alloc` dependency).

The suggested replacements are using `PasswordHash` if the goal was to
have an owned password hash type, or `String` if the goal is to have a
string representation of a PHC hash.