mcf: remove lifetime from `PasswordHashRef`
NOTE: breaking change

Changes the `PasswordHashRef` type into a `repr(transparent)` newtype
for `str`, constructing it as `&PasswordHashRef` using an unsafe pointer
cast which is only invoked after first validating it's well-formed.

This makes it possible to impl `Borrow<PasswordHashRef>` on
`PasswordHash` and `ToOwned` on `PasswordHashRef`.

`AsRef` and `Deref` impls have been added to `PasswordHash` for
obtaining a `&PasswordHashRef`, and the inherent methods which were
duplicated on both `PasswordHashRef` and `PasswordHash` have been
removed, following a similar pattern to `str`/`String`.

The `PasswordHash::as_mcf_hash_ref` method has been removed in favor of
using any of `AsRef`, `Borrow`, or `Deref` to do the conversion.