base64ct: deprecate `Base64Crypt` (#2135)
Closes #2133

`Base64ShaCrypt` is, as far as I can tell, the standard Base64 alphabet
used by all `crypt(3)` hashes. The only exception is `bcrypt` which uses
a completely different alphabet (`./A-Za-z0-9`).

As far as I can tell all algorithms that actually use the `./0-9A-Za-z`
alphabet do so with the little endian packing implemented in
`Base64ShaCrypt`. I'm uncertain about the provenance of the test vectors
for `Base64Crypt` but as far as I can tell it's actually a non-standard
packing for this particular alphabet.

As such, this deprecates `Base64Crypt`, steering people towards
`Base64ShaCrypt` for now. Perhaps in the future we can find better
names.